[refactor] Preview: move to utils package
